refactor: expose shared recipe builders from components
