+++ nemo_automodel/components/distributed/__init__.py
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from nemo_automodel.components.distributed.build import build_distributed
 from nemo_automodel.components.distributed.config import DDPConfig, FSDP2Config, MegatronFSDPConfig
 from nemo_automodel.components.distributed.pipelining.config import PipelineConfig
 
-__all__ = ["FSDP2Config", "MegatronFSDPConfig", "DDPConfig", "PipelineConfig"]
+__all__ = ["FSDP2Config", "MegatronFSDPConfig", "DDPConfig", "PipelineConfig", "build_distributed"]
