+++ nemo_automodel/components/distributed/build.py
@@ -1,0 +1,36 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
+from nemo_automodel.components.distributed.init_utils import initialize_distributed
+
+if TYPE_CHECKING:
+    from nemo_automodel.components.distributed.init_utils import DistInfo
+
+
+def build_distributed(cfg_dist: dict[str, Any]) -> "DistInfo":
+    """Build and initialize distributed training resources.
+
+    Args:
+        cfg_dist: Configuration for distributed training.
+
+    Returns:
+        Distributed training information from initialize_distributed.
+    """
+    backend = cfg_dist.get("backend", "nccl")
+    timeout = cfg_dist.get("timeout_minutes", 1)
+    return initialize_distributed(backend=backend, timeout_minutes=timeout)
