+++ nemo_automodel/components/loss/__init__.py
@@ -11,3 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+from nemo_automodel.components.loss.build import build_loss_fn
+
+__all__ = ["build_loss_fn"]
