+++ nemo_automodel/recipes/llm/train_seq_cls.py
@@ -23,10 +23,14 @@ import wandb
 
 from nemo_automodel._transformers.mfu import AutoMFU
 from nemo_automodel._transformers.utils import apply_cache_compatibility_patches
+from nemo_automodel.components.checkpoint import build_checkpoint_config
 from nemo_automodel.components.config._arg_parser import parse_args_and_load_config
+from nemo_automodel.components.distributed import build_distributed
 from nemo_automodel.components.loggers.log_utils import setup_logging
 from nemo_automodel.components.loggers.metric_logger import MetricsSample, build_metric_logger
-from nemo_automodel.components.loggers.wandb_utils import suppress_wandb_log_messages
+from nemo_automodel.components.loggers.wandb_utils import build_wandb, suppress_wandb_log_messages
+from nemo_automodel.components.optim import build_lr_scheduler
+from nemo_automodel.components.training import build_step_scheduler
 from nemo_automodel.components.training.rng import StatefulRNG
 from nemo_automodel.components.training.utils import clip_grad_norm
 from nemo_automodel.components.utils.flops_utils import calculate_mfu
@@ -35,13 +39,9 @@ from nemo_automodel.recipes._dist_setup import setup_distributed
 from nemo_automodel.recipes.base_recipe import BaseRecipe
 from nemo_automodel.recipes.llm.train_ft import (
     _get_model_name,
-    build_checkpoint_config,
     build_dataloader,
-    build_distributed,
-    build_lr_scheduler,
     build_model,
     build_optimizer,
-    build_step_scheduler,
 )
 
 logger = logging.getLogger(__name__)
@@ -69,9 +69,6 @@ class TrainFinetuneRecipeForSequenceClassification(BaseRecipe):
 
         if self.dist_env.is_main and hasattr(self.cfg, "wandb"):
             suppress_wandb_log_messages()
-            # Reuse helper from NTP recipe
-            from nemo_automodel.recipes.llm.train_ft import build_wandb
-
             run = build_wandb(self.cfg)
             logging.info("🚀 View run at {}".format(run.url))
