+++ nemo_automodel/recipes/retrieval/train_bi_encoder.py
@@ -26,23 +26,20 @@ from torch.utils.data import IterableDataset
 from torchdata.stateful_dataloader.sampler import StatefulDistributedSampler
 
 from nemo_automodel._transformers.utils import apply_cache_compatibility_patches
+from nemo_automodel.components.checkpoint import build_checkpoint_config
 from nemo_automodel.components.checkpoint.checkpointing import Checkpointer
 from nemo_automodel.components.config._arg_parser import parse_args_and_load_config
+from nemo_automodel.components.distributed import build_distributed
 from nemo_automodel.components.distributed.utils import FirstRankPerNode, get_sync_ctx
 from nemo_automodel.components.loggers.log_utils import setup_logging
 from nemo_automodel.components.loggers.metric_logger import MetricsSample, build_metric_logger
-from nemo_automodel.components.loggers.wandb_utils import suppress_wandb_log_messages
+from nemo_automodel.components.loggers.wandb_utils import build_wandb, suppress_wandb_log_messages
+from nemo_automodel.components.optim import build_lr_scheduler
+from nemo_automodel.components.training import build_step_scheduler
 from nemo_automodel.components.training.rng import ScopedRNG, StatefulRNG
 from nemo_automodel.components.training.utils import scale_grads_and_clip_grad_norm
 from nemo_automodel.recipes._dist_setup import setup_distributed
 from nemo_automodel.recipes.base_recipe import BaseRecipe
-from nemo_automodel.recipes.llm.train_ft import (
-    build_checkpoint_config,
-    build_distributed,
-    build_lr_scheduler,
-    build_step_scheduler,
-    build_wandb,
-)
 from nemo_automodel.shared.te_patches import apply_te_patches
 
 logger = logging.getLogger(__name__)
