+++ tests/unit_tests/recipes/test_train_ft.py
@@ -305,7 +305,7 @@ def test_build_checkpoint_config_peft_torch_save_overrides_to_safetensors(caplog
         "is_async": True,
     }
 
-    with caplog.at_level(logging.WARNING, logger="nemo_automodel.recipes.llm.train_ft"):
+    with caplog.at_level(logging.WARNING, logger="nemo_automodel.components.checkpoint.config_builder"):
         config = build_checkpoint_config(
             cfg_ckpt=cfg_ckpt,
             cache_dir=None,
